feat: 포인트 도메인 작업
