--- src/main/java/kr/hhplus/be/server/point/domain/Point.java
@@ -1,43 +1,0 @@
-package kr.hhplus.be.server.point.domain;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-import kr.hhplus.be.server.common.BaseEntity;
-import kr.hhplus.be.server.user.domain.User;
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@Table(name = "point")
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Point extends BaseEntity {
-
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	@Column(name = "point_id")
-	private Long pointId;
-
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "user_id")
-	private User user;
-
-	@Column
-	private Integer balance;
-
-	@Builder
-	public Point(Long pointId, User user, Integer balance) {
-		this.pointId = pointId;
-		this.user = user;
-		this.balance = balance;
-	}
-}
