+++ src/main/java/kr/hhplus/be/server/point/domain/entity/Point.java
@@ -1,0 +1,33 @@
+package kr.hhplus.be.server.point.domain.entity;
+
+import jakarta.persistence.*;
+import kr.hhplus.be.server.common.BaseEntity;
+import lombok.AccessLevel;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@Table(name = "point")
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Point extends BaseEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "point_id")
+    private Long pointId;
+
+    @Column(name = "user_id")
+    private Long userId;
+
+    @Column
+    private Integer balance;
+
+    @Builder
+    public Point(Long pointId, Long userId, Integer balance) {
+        this.pointId = pointId;
+        this.userId = userId;
+        this.balance = balance;
+    }
+}
