+++ src/main/java/kr/hhplus/be/server/point/domain/entity/PointTypeEnum.java
@@ -1,0 +1,13 @@
+package kr.hhplus.be.server.point.domain.entity;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public enum PointTypeEnum {
+    CHARGE("충전"),
+    USE("사용");
+
+    private final String text;
+}
