+++ src/main/java/kr/hhplus/be/server/point/domain/repository/PointRepository.java
@@ -1,0 +1,4 @@
+package kr.hhplus.be.server.point.domain.repository;
+
+public interface PointRepository {
+}
