+++ src/main/java/kr/hhplus/be/server/point/domain/service/PointService.java
@@ -1,0 +1,10 @@
+package kr.hhplus.be.server.point.domain.service;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class PointService {
+
+}
