--- src/main/java/kr/hhplus/be/server/point/infrastructure/PointHistoryRepository.java
@@ -1,8 +1,0 @@
-package kr.hhplus.be.server.point.infrastructure;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import kr.hhplus.be.server.point.domain.PointHistory;
-
-public interface PointHistoryRepository extends JpaRepository<PointHistory, Long> {
-}
