+++ src/main/java/kr/hhplus/be/server/point/infrastructure/PointHistoryRepositoryImpl.java
@@ -1,0 +1,6 @@
+package kr.hhplus.be.server.point.infrastructure;
+
+import kr.hhplus.be.server.point.domain.repository.PointHistoryRepository;
+
+public class PointHistoryRepositoryImpl implements PointHistoryRepository {
+}
