+++ src/main/java/kr/hhplus/be/server/point/infrastructure/PointRepositoryImpl.java
@@ -1,0 +1,6 @@
+package kr.hhplus.be.server.point.infrastructure;
+
+import kr.hhplus.be.server.point.domain.repository.PointRepository;
+
+public class PointRepositoryImpl implements PointRepository {
+}
