+++ src/main/java/kr/hhplus/be/server/seat/domain/Seat.java
@@ -1,16 +1,8 @@
 package kr.hhplus.be.server.seat.domain;
 
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-import kr.hhplus.be.server.concert.domain.entity.Concert;
+import jakarta.persistence.*;
 import kr.hhplus.be.server.common.BaseEntity;
+import kr.hhplus.be.server.concert.domain.entity.Concert;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
@@ -22,22 +14,22 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class Seat extends BaseEntity {
 
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	@Column(name = "seat_id")
-	private Long seatId;
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "seat_id")
+    private Long seatId;
 
-	@Column(name = "seat_number")
-	private Integer seatNumber;
+    @Column(name = "seat_number")
+    private Integer seatNumber;
 
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "concert_id")
-	private Concert concert;
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "concert_id")
+    private Concert concert;
 
-	@Builder
-	public Seat(Long seatId, Integer seatNumber, Concert concert) {
-		this.seatId = seatId;
-		this.seatNumber = seatNumber;
-		this.concert = concert;
-	}
+    @Builder
+    public Seat(Long seatId, Integer seatNumber, Concert concert) {
+        this.seatId = seatId;
+        this.seatNumber = seatNumber;
+        this.concert = concert;
+    }
 }
